WGJ-40 Fixed bug after changes to saved searches breaking searchbar functionality
After maps API implementation, a change was made to savedsearches.py that checked any city name with a comma as coordinates. With suggested searches, city names come with a city name, state name, and country separated by commas. We already had a coordinates to name backend conversion route. So I removed the code that was breaking everything and just used the backend route to convert the coordinates before making the backend call.